Fixed the issue of jupyter not respecting the virtualenv.
The jupyter was installed system wide and it was only using the system
python paths ignoring the virtualenv of the application. Fixed it by
installing jupyter in the virtualenv. The same applies for firefly as
well.

Also adding the virtualenv path to PATH env variable to make it the
default one.